Add the ability to fzf a README via GitHub stars
This PR, is more or less a rewrite that adds 3 new commands:

* `FetchStars`: uses [octokit](https://github.com/octokit/octokit.rb) to
store a user's starred repositories in `starred_repositories.txt`.
* `FindREADME`: uses fzf to search the contents of the
`starred_repositories.txt` file and navigates to the `#readme` css
selector on the repository's homepage. If the user has not populated
their starred repositories, they will be prompted to run the
`FetchStars` command.
* `OpenREADME`: allows a user to open any repository, starred or not,
via `:OpenREADME anhari/vim-readme`, for example.

Also updates the documenation, and fixes #1 and #2.